# swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/reminder": {
      "post": {
        "operationId": "ReminderController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReminderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The reminder has been successfully created."
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new reminder",
        "tags": [
          "Reminders"
        ]
      },
      "get": {
        "operationId": "ReminderController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "showCompleted",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return all reminders for the user."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all reminders for the logged-in user",
        "tags": [
          "Reminders"
        ]
      }
    },
    "/api/v1/reminder/{id}": {
      "get": {
        "operationId": "ReminderController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the reminder."
          },
          "404": {
            "description": "Reminder not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a reminder by id",
        "tags": [
          "Reminders"
        ]
      },
      "put": {
        "operationId": "ReminderController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReminderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The reminder has been successfully updated."
          },
          "404": {
            "description": "Reminder not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a reminder",
        "tags": [
          "Reminders"
        ]
      },
      "delete": {
        "operationId": "ReminderController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The reminder has been successfully deleted."
          },
          "404": {
            "description": "Reminder not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a reminder",
        "tags": [
          "Reminders"
        ]
      }
    },
    "/api/v1/reminder/completed": {
      "delete": {
        "operationId": "ReminderController_deleteAllCompleted",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Completed reminders have been successfully deleted."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete all completed reminders for the logged-in user",
        "tags": [
          "Reminders"
        ]
      }
    },
...
    "/api/v1/ai/inspiration": {
      "post": {
        "operationId": "AiController_getInspirationalMessage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "longitude": {
                    "type": "number"
                  },
                  "latitude": {
                    "type": "number"
                  }
                },
                "required": [
                  "longitude",
                  "latitude"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns an inspirational message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get inspirational message based on user context",
        "tags": [
          "AI"
        ]
      }
    },
    "/api/v1/note/{reminderId}/notes": {
      "post": {
        "operationId": "NoteController_createNotes",
        "parameters": [
          {
            "name": "reminderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The note has been successfully created for the reminder."
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new note for the reminder",
        "tags": [
          "Notes"
        ]
      },
      "get": {
        "operationId": "NoteController_findNotes",
        "parameters": [
          {
            "name": "reminderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The note has been successfully fetched for the reminder. "
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Find an existing notes for a reminder",
        "tags": [
          "Notes"
        ]
      }
    },
    "/api/v1/note/{reminderId}/notes/{noteId}": {
      "put": {
        "operationId": "NoteController_updateNote",
        "parameters": [
          {
            "name": "reminderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "noteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The note has been successfully updated for the reminder. "
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an existing note for a reminder",
        "tags": [
          "Notes"
        ]
      },
      "get": {
        "operationId": "NoteController_findNote",
        "parameters": [
          {
            "name": "reminderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "noteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The note has been successfully fetched for the reminder. "
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Find an existing notes for a reminder",
        "tags": [
          "Notes"
        ]
      },
      "delete": {
        "operationId": "NoteController_deleteNote",
        "parameters": [
          {
            "name": "reminderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "noteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The note has been successfully deleted for the reminder. "
          },
          "400": {
            "description": "Bad Request."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an existing note for a reminder",
        "tags": [
          "Notes"
        ]
      }
    }
  },
  "info": {
    "title": "Reminder App API",
    "description": "The Reminder App API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateReminderDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "example": "Meeting at 10 AM",
            "description": "The description of the reminder or todo"
          },
          "dueDate": {
            "format": "date-time",
            "type": "string",
            "example": "2022-01-15T10:00:00.000Z",
            "description": "The date and time when the reminder or todo is due"
          },
          "status": {
            "type": "boolean",
            "example": true,
            "description": "The status of the reminder or todo (true for completed, false for not completed)"
          },
          "priority": {
            "type": "boolean",
            "example": true,
            "description": "Indicates this reminder is a priority"
          }
        },
        "required": [
          "description",
          "dueDate",
          "status",
          "priority"
        ]
      },
      "UpdateReminderDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "example": "Meeting at 10 AM",
            "description": "The description of the reminder or todo"
          },
          "dueDate": {
            "format": "date-time",
            "type": "string",
            "example": "2022-01-15T10:00:00.000Z",
            "description": "The date and time when the reminder or todo is due"
          },
          "status": {
            "type": "boolean",
            "example": true,
            "description": "The status of the reminder or todo (true for completed, false for not completed)"
          },
          "priority": {
            "type": "boolean",
            "example": true,
            "description": "Indicates this reminder is a priority"
          },
          "notes": {
            "description": "Array of note IDs to set for this reminder",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateNoteDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "I need to buy groceries",
            "description": "The note that should be added to the reminders"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateNoteDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "I need to buy groceries",
            "description": "The note that should be added to the reminders"
          }
        }
      }
    }
  }
}